rewrite jack audio driver interface
It avoids the synchronization between c and lisp during the
processing. The lisp thread controls directly the jack client,
and the c thread is used only when lisp is busy.

The buffer size is updated in the realtime thread at the start of
a new cycle.

TODO: stop the lisp rt thread if the jack server kills the client