inference of the performance-time variables in a VUG-MACRO
Not all arguments of a VUG-MACRO are the inputs of the VUG,
therefore we have to explicitly specify this information by
using the VUG-INPUT function or the new macro WITH-VUG-INPUTS,
otherwise the inference of the performance-time variables will
fail. The next example is ok:

(dsp! vug-input-test ()
  (stereo (impulse (line 8 1 3 #'free) .7)))

but the follow fails if there isn't the info about the inputs of the
VUG in the definition of IMPULSE VUG-MACRO:

(dsp! vug-input-test ()
  (with-samples ((f 0.0))
    (foreach-tick (setf f (line 8 1 3 #'free)))
    (stereo (impulse f .7))))

All the built-in VUG-MACROs are fixed.