temporal envelope for the read-macro #[... beat ...]
added also the new utility MAKE-TEMPO-ENVELOPE to create an
envelope where the input levels are the values of the tempo in
bpm, but the levels of the generated envelope are in bps.

Example:

SCRATCH> *tempo*
SCRATCH> *sample-rate*
48000.0d0
SCRATCH> #[1 b]
48000.0d0
SCRATCH> #[1 b *tempo*]
48000.0d0

If we pass the fourth argument, the third is an envelope where
the levels are the values of the tempo in bps.

;; Acceleration from 60 to 120 bpm in 16 beats with an exponential curve.
(defvar *tempo-env* (make-tempo-envelope '(60 120) '(16) :curve -4))

SCRATCH> #[1 b *tempo-env* 0]
48000.0d0
SCRATCH> #[1 b *tempo-env* 4]
32546.057762688662d0
SCRATCH> #[1 b *tempo-env* 8]
26860.870128530823d0
SCRATCH> #[1 b *tempo-env* 16]
24000.0d0
SCRATCH> #[1 b *tempo-env* 20]
24000.0d0