new parser for the score file
A score file can contain time-tagged lisp functions, lisp statements
and lisp tags.

A line comment starts with `;' and there is a line continuation with
`\' at the end of the line.

TIME and TEMPO-ENV are pre-defined local variables usable inside a
score file. They are respectively the initial time in samples and the
temporal envelope of the events.

It is possible to define other local variables by inserting the
bindings after WITH, at the beginning of the score. We can also add a
DECLARE expression after the bindings.

DUR is a local function to convert the duration from beats to seconds
with respect to TEMPO-ENV.

TEMPO is a local macro to change the tempo of the score.