incudine is also the name of a standalone executable
`make install' or `make HAVE_LV2=1 install' to install it.

The location of the command is `/usr/local/bin' by default but
we can change the variables PREFIX and/or BINDIR.

The variables DYNAMIC_SPACE_SIZE and CONTROL_STACK_SIZE are
useful to set respectively the runtime options
--dynamic-space-size and --control-stack-size of SBCL (all the
arguments of the "incudine" command are passed to the toplevel).