EDF scheduling: truncate the fractional value of the time.
It is possible to sort the events scheduled at the same time by using
the fractional part of the time. For example:

(let ((t1 (+ (now) *sample-rate*)))
  (at (+ t1 0.1) #'f1)
  (at (+ t1 0.2) #'f2)
  (at (+ t1 0.3) #'f3))

the sequence of the functions F1, F2 and F3 is called at the time T1.
Without the fractional offset, the order of the functions is undefined.