cudere-clm: new version of SRC
The conversion is performed through the convolution between the input,
upsampled by rate, and the polyphase filters obtained from the windowed
sinc table.

The aliasing is remarkably reduced and the code is simplified.

The delay between the polyphase filters is not one sample, but it
depends on the rate.

Also added a new argument WINDOW-BETA to MAKE-SRC, to set the
parameter of the Kaiser window. If WINDOW-BETA is NIL, the sinc
table is windowed by a Hanning window (the default also in the
original implementation).