cudere-clm: the new version of SRC is optional during the development.
Currently, pre-filtering of the input is sometimes required if
the rate is greater than one. It is to fix. In practice, the
current filter is good for decimation, but the upsampling requires
another filter (or other polyphase filters from the same table).

Use MAKE-SRC with SRC-TYPE EXPERIMENTAL to test the new version.