WITH-SCHEDULE sets the local time to zero.
It also works with the code that depends on INCUDINE:NOW,
for example in Common Music if the method CM:WRITE-EVENT
schedules at INCUDINE:NOW plus scoretime:

    (with-schedule
      (events (schedule-many-events) rt-stream))

Memo: the configuration variable *RT-EDF-HEAP-SIZE* in .incudinerc
sets the max number of scheduled events.

The sample counter is zero and read-only inside WITH-SCHEDULE:

    (with-schedule (incf (now)) (princ (now)))
    ;; => 0.0d0

However, we can require a SETFable local counter:

    (with-schedule (with-local-time () (incf (now)) (princ (now))))
    ;; => 1.0d0

Note: all the events are scheduled with the same time offset during
the copy from the temporary queue to the real-time queue. That time
is the value of NOW from the real-time sample counter. In practice,
the optional local counter from WITH-LOCAL-TIME is another time offset
in samples added through NOW. For example, if the scheduling depends
on INCUDINE:NOW, a delay of one second is:

    (with-schedule (with-local-time (:start #[1 s]) (princ (now))))
    ;; => 48000.0d0