regofile: experimental call/return statements.
The score statement `call' pushes the return position on the stack
and transfers program control to the point labeled by a tag.
The score statement `return' transfers control to the return position
located on the top of the stack.

Syntax for call statement (a tag between [[ ]] is a facility for Org mode):

    call tag
    call tag time
    call [[tag]] time                ; the target label is <<tag>>
    call [[tag][description]] time   ; the target label is <<tag>>

Example:

    call p1 0
    call [[p2][pattern two]] 1
    call p3 1.5
    call p1 2
    return              ; end of score

    p1
    0 write-line "pattern 1" // force-output
    call p3 .1
    call p3 .25
    return

    <<p2>>
    0 write-line "pattern 2" // force-output
    return

    p3
    0 write-line "pattern 3" // force-output
    return