Reorganize the test for the required C libraries and header files.
In particular, X11 is used only for mouse support, therefore the
compilation continues with a verbose warning if the header files
are not correctly installed. On the contrary, the compilation
terminates with an error if the required libraries or header files
for real time audio support are not found (i.e. the configuration
variable *AUDIO-DRIVER* is set to :JACK or :PORTAUDIO but the linker
fails during the test).