Add the specification :START-TIME to WITH-SCHEDULE.
The temporary queue introduces a generally unspecified latency.
If the first form in BODY is a list

    (:start-time value-form)

VALUE-FORM is the absolute time offset in samples of the first
scheduled event if

    (>= value-form (+ (now) latency))

If it is necessary to change the start-time within BODY, the
specification is

    (:start-time initial-value-form variable)

In this case, a VARIABLE is bound to a time-getter function. It is a
function because the computation occurs in rt-thread.

Examples:

    (with-schedule
      (:start-time (tempo-sync #[4 beats]))
      ...)

    (with-schedule
      (:start-time (incudine.edf:next-time #'phrase-7))
      ...)

    (with-schedule
      (:start-time 0 start)
      ;; Initially, the variable START is `(lambda () 0)'
      ...
      (setf start (lambda () (tempo-sync #[1 b])))
      ...)