windows: rt-thread associated with "Pro Audio" task if MMCSS is available.
If necessary, we can add a task ("Audio" or other) for fast-nrt-thread
or nrt-thread calling ADD-THREAD-TASK with FAST-NRT-FUNCALL or NRT-FUNCALL
(the calling thread is associated with the specified task).

Note: by default, "Pro Audio" priority is higher than "Audio" priority.

Currently ADD-THREAD-TASK and REMOVE-THREAD-TASK are not exported,
and the other priority settings are ignored on Windows.