OSC:RECEIVE reads the OSC messages from the received OSC bundle too.
The current OSC message is the first message of the received OSC bundle.
The other messages are queued for the successive calls to OSC:RECEIVE.

A responder for a OSC:INPUT-STREAM works with the current OSC message,
therefore an OSC bundle doesn't require a specific interface.

OSC:MESSAGE-TIME returns the time obtained from the OSC time tag, or
zero if the received OSC packet is not an OSC bundle. By default the
time value is suitable as first argument to INCUDINE:AT

    (at (osc:message-time stream) function ...)

where message-time is the OSC bundle time or zero.

The alternative time unit for OSC:MESSAGE-TIME is SECONDS in universal
time format.