Add OSC:END-OF-BUNDLE-P.
The secondary value is T if the current message is part of a OSC bundle.

A hypothetical utility OSC:BUNDLE-P, with the reversed returned values,
is generally useless with an OSC output stream because we write the
contents of the OSC packet. OSC:END-OF-BUNDLE-P is more important for an
OSC input stream because the messages are queued for the successive
calls to OSC:RECEIVE. Moreover, a received packet is a bundle if
OSC:MESSAGE-TIME is non zero.